src/net/sourceforge/mayfly/evaluation/condition/And.java
package net.sourceforge.mayfly.evaluation.condition;

import net.sourceforge.mayfly.evaluation.ResultRow;


public class And extends BooleanExpression {
    public final BooleanExpression leftSide;
    public final BooleanExpression rightSide;

    public And(BooleanExpression leftSide, BooleanExpression rightSide) {
        this.leftSide = leftSide;
        this.rightSide = rightSide;
    }

    public boolean evaluate(ResultRow row) {
        return leftSide.evaluate(row) && rightSide.evaluate(row);
    }

    public String firstAggregate() {
        return firstAggregate(leftSide, rightSide);
    }

}
